there may not be an embed

+++ cogs/music/_musicplayer.py
@@ -368,7 +368,8 @@ class MusicPlayer:
         self.logger.debug("movehere command")
 
         # Delete the old message
-        await self.embed.delete()
+        if self.embed:
+            await self.embed.delete()
         # Set the channel to this channel
         self.embed.channel = channel
         # Send a new embed to the channel
